backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==23.5
# psycopg2-binary==2.9.9
Pillow
python-dotenv==1.0.0
drf-spectacular==0.27.1
Faker==24.0.0
